requirements.txt
transformers
pillow
torch
requests
pytest
timm